add s3 fetcher, and a test

+++ tika-pipes-grpc/docker-build/docker-build.sh
@@ -44,14 +44,17 @@ cd "${OUT_DIR}" || exit
 
 echo "Running docker build from directory: $(pwd)"
 
-# build single arch
-#docker build . -t "${TAG_NAME}"
-
-docker buildx create --name tikapipesbuilder
-# see https://askubuntu.com/questions/1339558/cant-build-dockerfile-for-arm64-due-to-libc-bin-segmentation-fault/1398147#1398147
-docker run --rm --privileged tonistiigi/binfmt --install amd64
-docker run --rm --privileged tonistiigi/binfmt --install arm64
-docker buildx build --builder=tikapipesbuilder . -t "${TAG_NAME}" --platform linux/amd64,linux/arm64 --push
-docker buildx stop tikapipesbuilder
+if [ -z "${MULTI_ARCH}" ]; then
+  # build single arch
+  docker build . -t "${TAG_NAME}"
+else
+  # build multi arch
+  docker buildx create --name tikapipesbuilder
+  # see https://askubuntu.com/questions/1339558/cant-build-dockerfile-for-arm64-due-to-libc-bin-segmentation-fault/1398147#1398147
+  docker run --rm --privileged tonistiigi/binfmt --install amd64
+  docker run --rm --privileged tonistiigi/binfmt --install arm64
+  docker buildx build --builder=tikapipesbuilder . -t "${TAG_NAME}" --platform linux/amd64,linux/arm64 --push
+  docker buildx stop tikapipesbuilder
+fi
 
 echo "Done running docker build for tag ${TAG_NAME}"
